Refactor appointmentNote integration tests to use factories and cleanup utilities
- Remove eslint-disable and split large test file into focused test files per HTTP method
- Replace manual test data creation with factory functions from @mcw/database/mock-data
- Use centralized cleanupDatabase() utility instead of manual cleanup code
- Update CLAUDE.md with factory usage patterns and file/method size guidelines
- Create comprehensive TESTING_GUIDELINES.md documentation
- Each test file now under 250 lines (was 422 lines)

Addresses PR #126 feedback from talhabalaj about using mock data creators and avoiding eslint-disable
