+++ apps/back-office/__tests__/api/appointmentNote/route.put.integration.test.ts
@@ -1,0 +1,244 @@
+import { describe, it, expect, beforeEach, afterEach, vi } from "vitest";
+import { PUT } from "@/api/appointmentNote/route";
+import {
+  prisma,
+  Practice,
+  User,
+  Client,
+  Clinician,
+  SurveyTemplate,
+  Appointment,
+  SurveyAnswers,
+} from "@mcw/database";
+import {
+  UserFactory,
+  ClientFactory,
+  ClinicianPrismaFactory,
+  SurveyTemplatePrismaFactory,
+  AppointmentPrismaFactory,
+  SurveyAnswersPrismaFactory,
+} from "@mcw/database/mock-data";
+import { cleanupDatabase } from "@mcw/database/test-utils";
+import { createRequestWithBody } from "@mcw/utils";
+import { getBackOfficeSession } from "@/utils/helpers";
+import type { Session } from "next-auth";
+
+vi.mock("@/utils/helpers", () => ({
+  getBackOfficeSession: vi.fn(),
+}));
+
+const mockSession = {
+  user: {
+    id: "test-user-id",
+    email: "test@example.com",
+    practice_id: "test-practice-id",
+    role: "ADMIN",
+  },
+};
+
+describe("appointmentNote API - PUT Integration Tests", () => {
+  let testPractice: Practice;
+  let testUser: User;
+  let testClient: Client;
+  let testClinician: Clinician;
+  let testTemplate: SurveyTemplate;
+  let testAppointment: Appointment;
+  let existingNote: SurveyAnswers;
+
+  beforeEach(async () => {
+    vi.mocked(getBackOfficeSession).mockResolvedValue(mockSession as Session);
+
+    // Create test practice first (no factory available)
+    testPractice = await prisma.practice.create({
+      data: {
+        name: "Test Practice",
+        email: "practice@test.com",
+        phone: "1234567890",
+        city: "Test City",
+        state: "TS",
+        zip: "12345",
+        address_line_1: "123 Test St",
+      },
+    });
+
+    // Create test user using factory
+    testUser = await prisma.user.create({
+      data: {
+        ...UserFactory.build({
+          email: "test@example.com",
+          role: "ADMIN",
+        }),
+        practice_id: testPractice.id,
+      },
+    });
+
+    // Create test client using factory
+    testClient = await prisma.client.create({
+      data: {
+        ...ClientFactory.build({
+          email: "john.doe@example.com",
+        }),
+        practice_id: testPractice.id,
+      },
+    });
+
+    // Create test clinician using factory
+    testClinician = await ClinicianPrismaFactory.create({
+      user_id: testUser.id,
+      practice_id: testPractice.id,
+    });
+
+    // Create test template using factory
+    testTemplate = await SurveyTemplatePrismaFactory.create({
+      name: "Progress Note Template",
+      type: "PROGRESS_NOTES",
+      content: JSON.stringify({
+        pages: [
+          {
+            elements: [
+              {
+                type: "text",
+                name: "question1",
+                title: "How are you feeling today?",
+              },
+            ],
+          },
+        ],
+      }),
+      is_active: true,
+      practice_id: testPractice.id,
+    });
+
+    // Create test appointment using factory
+    testAppointment = await AppointmentPrismaFactory.create({
+      title: "Test Appointment",
+      status: "SCHEDULED",
+      appointment_type_id: "90834",
+      clinician_id: testClinician.id,
+      practice_id: testPractice.id,
+    });
+
+    // Create appointment-client relationship
+    await prisma.appointmentClients.create({
+      data: {
+        appointment_id: testAppointment.id,
+        client_id: testClient.id,
+      },
+    });
+
+    // Create existing note for update tests
+    existingNote = await SurveyAnswersPrismaFactory.create({
+      template_id: testTemplate.id,
+      client_id: testClient.id,
+      content: JSON.stringify({ question1: "Original answer" }),
+      status: "DRAFT",
+      appointment_id: testAppointment.id,
+    });
+  });
+
+  afterEach(async () => {
+    // Use centralized cleanup utility
+    await cleanupDatabase(prisma, { verbose: false });
+  });
+
+  it("should update existing note by appointment_id", async () => {
+    const updatePayload = {
+      id: testAppointment.id,
+      content: JSON.stringify({ question1: "Updated answer" }),
+      status: "COMPLETED",
+    };
+
+    const request = createRequestWithBody(
+      "/api/appointmentNote",
+      updatePayload,
+      "PUT",
+    );
+    const response = await PUT(request);
+
+    expect(response.status).toBe(200);
+    const data = await response.json();
+    expect(data.content).toBe(updatePayload.content);
+    expect(data.status).toBe("COMPLETED");
+
+    // Verify in database
+    const updated = await prisma.surveyAnswers.findUnique({
+      where: { id: existingNote.id },
+    });
+    expect(updated?.content).toBe(updatePayload.content);
+    expect(updated?.status).toBe("COMPLETED");
+  });
+
+  it("should update existing note by id", async () => {
+    const updatePayload = {
+      id: existingNote.id,
+      content: JSON.stringify({ question1: "Updated by ID" }),
+    };
+
+    const request = createRequestWithBody(
+      "/api/appointmentNote",
+      updatePayload,
+      "PUT",
+    );
+    const response = await PUT(request);
+
+    expect(response.status).toBe(200);
+    const data = await response.json();
+    expect(data.content).toBe(updatePayload.content);
+  });
+
+  it("should return 404 if note not found", async () => {
+    const updatePayload = {
+      id: "nonexistent-id",
+      content: JSON.stringify({ question1: "Updated" }),
+    };
+
+    const request = createRequestWithBody(
+      "/api/appointmentNote",
+      updatePayload,
+      "PUT",
+    );
+    const response = await PUT(request);
+
+    expect(response.status).toBe(404);
+    const data = await response.json();
+    expect(data.error).toBe("Note not found");
+  });
+
+  it("should return 401 for unauthenticated requests", async () => {
+    vi.mocked(getBackOfficeSession).mockResolvedValueOnce(null);
+
+    const updatePayload = {
+      id: existingNote.id,
+      content: JSON.stringify({ question1: "Updated" }),
+    };
+
+    const request = createRequestWithBody(
+      "/api/appointmentNote",
+      updatePayload,
+      "PUT",
+    );
+    const response = await PUT(request);
+
+    expect(response.status).toBe(401);
+    const data = await response.json();
+    expect(data.error).toBe("Unauthorized");
+  });
+
+  it("should validate update payload", async () => {
+    const invalidPayload = {
+      // Missing id
+      content: JSON.stringify({ question1: "Updated" }),
+    };
+
+    const request = createRequestWithBody(
+      "/api/appointmentNote",
+      invalidPayload,
+      "PUT",
+    );
+    const response = await PUT(request);
+
+    expect(response.status).toBe(400);
+    const data = await response.json();
+    expect(data.error).toBe("Invalid input");
+  });
+});
